Fix move Points method to Shape
- isSamePoint
