+++ src/main/java/location/domain/Points.java
@@ -1,31 +1,19 @@
 package location.domain;
 
-import java.util.*;
+import java.util.List;
 
 public class Points {
     private final static String POINT_NULL_OR_BLANK_MSG = "포인트가 비어있습니다.";
-    private final static String DUPLICATE_MSG = "중복된 좌표가 있습니다.";
     private static List<Point> points;
 
     public Points(final List<Point> points) {
-        isValid(points);
+        checkValid(points);
         compare(points);
         Points.points = points;
     }
 
-    private void isValid(final List<Point> points) {
+    private void checkValid(final List<Point> points) {
         isNullOrBlank(points);
-        isSamePoint(points);
-    }
-
-    private void isSamePoint(final List<Point> points) {
-        Set<Point> notDuplicatePoints = new HashSet<>();
-        for (Point point : points) {
-            notDuplicatePoints.add(point);
-        }
-        if (notDuplicatePoints.size() != points.size()) {
-            throw new IllegalArgumentException(DUPLICATE_MSG);
-        }
     }
 
     private void isNullOrBlank(final List<Point> points) {
@@ -49,5 +37,9 @@ public class Points {
 
     public int size() {
         return points.size();
+    }
+
+    public List<Point> getPoints() {
+        return points;
     }
 }
